refactor: rename Interests to HobbyProjects
